backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
cloudinary
python-multipart
python-dotenv
pydantic
pydantic-settings
dependency-injector
